Move post-processor to separate module
